requirement.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.3
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.0
idna==2.7
jdcal==1.4
mysqlclient==1.3.10
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
pycparser==2.19
pycryptodome==3.7.2
pymongo==3.7.2
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0
